New main

See merge request javapro17/javapro-socialnetwork-studygroup-17!57

+++ src/main/java/ru/skillbox/socialnetwork/service/impl/PersonServiceImpl.java
@@ -14,9 +14,7 @@ import ru.skillbox.socialnetwork.data.dto.PersonResponse.Data;
 import ru.skillbox.socialnetwork.data.dto.PersonSearchResponse;
 import ru.skillbox.socialnetwork.data.entity.File;
 import ru.skillbox.socialnetwork.data.entity.Person;
-import ru.skillbox.socialnetwork.data.repository.PersonRepo;
-import ru.skillbox.socialnetwork.data.repository.FileRepository;
-import ru.skillbox.socialnetwork.data.repository.TownRepository;
+import ru.skillbox.socialnetwork.data.repository.*;
 import ru.skillbox.socialnetwork.exception.PersonNotAuthorized;
 import ru.skillbox.socialnetwork.exception.UnauthorizedException;
 import ru.skillbox.socialnetwork.service.PersonService;
@@ -52,9 +50,20 @@ public class PersonServiceImpl implements PersonService {
     }
 
     @Override
-    public PersonResponse deletePerson(Principal principal) {
+    public PersonResponse deletePerson(Boolean isHardDelete, Principal principal) {
         Person person = findPerson(principal);
-        personRepository.delete(person);
+
+        if(isHardDelete) {
+            personRepository.delete(person);
+        } else {
+            person.setFirstName("Deleted");
+            person.setLastName("User");
+            person.setPhoto("https://static.thenounproject.com/png/438810-200.png");
+            person.setApproved(false);
+            person.setBlocked(true);
+            personRepository.save(person);
+        }
+        SecurityContextHolder.clearContext();
         return createSmallPersonResponse();
     }
